feat : 주간 식단 평가 기능 추가
- 주간 메뉴 전체에 대해 평가 결과 안내하는 함수
- evaluateMenu.js 를 사용하여 각 메뉴에 대해 모두 평가
